+++ moor/CHANGELOG.md
@@ -3,7 +3,7 @@
 - New `DatabaseConnection.delayed` constructor to synchronously obtain a database connection
   that requires async setup. This can be useful when connecting to a `MoorIsolate`.
 - `VmDatabase`: Create directory of database file to avoid misuse errors from sqlite3.
-- Add `groupConcat` function to Dart api.
+- Add `groupConcat` and `coalesce` functions to the Dart query builder.
 
 ## 3.3.1